requirements.txt
beautifulsoup4==4.4.1
certifi==2017.4.17
chardet==3.0.4
Django==1.8
django-extensions==1.9.1
djangorestframework==3.7.0
html5lib==0.999
idna==2.5
IMDbPY==5.0
lxml==3.5.0
pytz==2014.10
rabbitadmin==1.2.1
requests==2.18.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.1.18
urllib3==1.21.1
virtualenv==15.1.0
